pyproject.toml
[project]
name = "slack-archive-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "slack-sdk>=3.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]

[tool.ruff]
exclude = ["venv", ".venv"]
# Same as Black.
line-length = 88
indent-width = 4